Make getDataFromTree more universal + fix bug with instance created more times than necessary (#1741)
* Make getDataFromTree more universal

Make getDataFromTree more universal. The 'query' in variables was renamed to 'promise' to be more generic. 
Error message was also made to be more generic.

The 'fetchRoot' argument was removed due to instance reuse from previous walkTree cycle. Also getChildContext and componentWillMount is not called more times on the same element type instance with fetchData function than was necessary, which caused unwanted side effects.

* Remove unnecessary element attribute

* Update getDataFromTree.ts

* Fix variable name error in previous commit