Fix an error when a refetchQuery wasn't defined (#1794)
When we are seeking to find a query to refetch. We must take care to return the operation, if it exists, else we must return the original string.
The code in apollo-client does not handle a query equal to "undefined" but takes care of the string if properly forwarded to. See ahttps://github.com/apollographql/apollo-client/blob/master/packages/apollo-client/src/core/QueryManager.ts l254.