Reimplement invariant(condition, message) using TypeScript.
The official Facebook invariant implementation supports several features
that we do not need in react-apollo, such as message string formatting,
and runtime verification that a message string was provided.

Using this TypeScript implementation of the invariant function will
enforce that we do not use these extra features, while also allowing a
simpler implementation.

In the future, when we provide a production bundle, we will want to strip
the error strings from invariant calls, like React does.

The new InvariantError class can be thrown after manually checking a
condition in cases where it's important for TypeScript to understand the
condition (and use it for type narrowing later in the code). We can strip
the message strings from new InvariantError("...") expressions, too.