Adjust `startPolling` test avoid using `jest.runTimersToTime`
This test was depending on polling to complete within a specific
time interval, but wasn't accommodating for extra processing time
that might be needed during that interval. Instead of relying on
a specific time frame, these changes adapt the test to finish
after 3 polling intervals have completed (and polling time has
been reduced to help this test finish in half the time that it
did previously).