Add test for Query bug which is caused by changing props (#1749) (#2718)
* Add test for Query bug which is caused by changing props (#1749)

* Skip the failing test (#1749)

* Enable failing test

* Leverage lastResult to limit unnecessary subscription re-creation

When the Observable subscription created in `startQuerySubcription`
receives an error, we only want to remove the old subscription,
and start a new subscription, when we don't have a
previous result stored. So if no previous result was received due to
problems fetching data, or the previous result has been forcefully
cleared out for some reason, we shoule re-create a new
subscription. Otherwise we might be unnecessarily starting a new
subscription the `Query` component isn't expecting, and isn't tying
into its update cycle properly.

* Changelog update