Remove tslint
There are some advantages to using a linter with Typescript, but
they aren't really something we're interested in using with
React Apollo. Since tslint is being deprecated and reborn as an
eslint add-on, we're going to stop using it altogether. We might
consider bringing eslint into the project at some point in the
future, but for now we're not interested in imposing additional
rules or guidelines on potential contributors.