Make sure `useSubscription` cleans up when done (#3171)
* Make sure `useSubscription` cleans up when done

The `useSubscription` Hook isn't properly restting/removing the
internal `ObservableQuery` instance it uses, when it's no longer
needed. This means that the subscription used by the hook is
never properly shut down, leading to unnecessary network
communication. This commit ensures that the `useSubscription`
internals are properly cleansed when the component wrapping
the hook is unmounted.

Fixes #3160.

* Bundlesize increase