Switch back to the mutation function in first tuple position
While the changes requested in #3189 and implemented in #3199
seemed like a good idea, we no longer think it makes sense for
React Apollo to align its hooks API with React's core hooks API.
In `useMutation`'s case, the mutation function will almost always
be called, whereas the results aren't necessarily used. For this
reason it makes sense to keep the mutation function in the first
position of the result tuple.